feat(mobile): export diagnostics
