+++ .changeset/export-mobile-diagnostics.md
@@ -1,0 +1,5 @@
+---
+default: minor
+---
+
+Export diagnostics from mobile settings, bundling redacted logs and system info into a shareable archive.
