+++ src-tauri/Cargo.toml
@@ -46,6 +46,8 @@ async-stream = "0.3"
 base64 = "0.22"
 tokio-util = { version = "0.7", features = ["codec"] }
 futures-util = "0.3"
+regex = "1"
+zip = { version = "=4.6.1", default-features = false, features = ["deflate-flate2-zlib-rs"] }
 
 tauri = { version = "2.11.5", default-features = false, features = [
   "tray-icon",
@@ -101,8 +103,6 @@ tauri-plugin-updater = { version = "2", optional = true }
 tauri-plugin-process = "2"
 tauri-plugin-window-state = "2.4.1"
 tauri-plugin-global-shortcut = "2"
-regex = "1"
-zip = { version = "=4.6.1", default-features = false, features = ["deflate-flate2-zlib-rs"] }
 
 [target.'cfg(any(target_os = "linux", target_os = "dragonfly", target_os = "freebsd", target_os = "openbsd", target_os = "netbsd"))'.dependencies]
 gtk = "0.18"
