+++ src/app/features/settings/settingsLink.ts
@@ -35,6 +35,7 @@ export const settingsLinkFocusIdsBySection: Record<SettingsSectionId, readonly s
     'encrypted-room-url-preview',
     'enter-for-newline',
     'error-reporting',
+    'export-diagnostics',
     'file-description-placement',
     'hide-member-events-read-only-rooms',
     'hide-membership-change',
@@ -50,7 +51,6 @@ export const settingsLinkFocusIdsBySection: Record<SettingsSectionId, readonly s
     'right-aligned-bubbles',
     'right-swipe-action',
     'session-replay',
-    'export-diagnostics',
     'show-hidden-events',
     'hidden-event-edits',
     'show-redacted-message-tombstones',
